Rename loginService as apiService
